2.10.0.1 Beta New detections are added:  - O4 - some new detections and improved fixes.  - O5 - Applet: to detect custom and hijacked control panel items.  - O7 - AppLocker:  	* NOTE!!! Applocker is by default in whitelisting mode: everything that is not explicitly allowed, is blocked.  	* If you remove "(allow)" rule, this particular application/folder will be blocked, until you'll remove all the rules.  	* To remove all rules at once, it's better to fix the special line "O7 - AppLocker: fix all"  - O7 - KnownFolder: to detect and fix hijacked locations of "Known Folders".  - O7 - TroubleShooting: new environment variables are added.  - O22 - BITS Job: to detect foreign Windows Update Service jobs.  - O22 - Tasks: (damage) subsection is added to detect damaged and garbage tasks of such types: 	* (user missing) - when user/group is removed; 	* (no xml) - when xml file doesn't exist; 	* (key missing) - when associated reg. key doesn't exist; 	* (no key) - when no task keys are referred to; 	* (empty) - when tasks folder or key contains an empty record; 	* This doesn't include xml integrity/CRC check and "DynamicInfo / Triggers" validation. Interface:  - Added ability to mark multiple items for fixing with 'Shift + Click'.  - Added "Copy" multi-context menu in scan results window (thanks @TheTrick for helping with Clipboard).  - Added "VirusTotal" multi-context menu in scan results window with options to check file/URL by hash or to submit file with SysInternals Autoruns. Report: > General:  - Sorting is now going in correct alphabetical order + increased the speed.  - O4 - Startup subsections are renamed to improve the sorting. > /Area:Environment:  - Added registry report of "User Shell Folders" and "Shell Folders" in addition to CLSID-based report.  Why? Because Microsoft doesn't follow its own 'Best Practice' from MSDN. lol. > Modules "Check Browsers' LNK" & "ClearLNK":  - Suppressed "Allow to download..." request when silent mode is selected in "Update" settings.  - Added auto-update feature (whenever you run the tool) - not often than once per month.  - Introduced the strong verification for the file digital signature before execution.  - Tools are now will be downloaded in \Tools\Scan subfolder of HiJackThis dir or installation dir if one is performed. Other:  - O4 - Fixed reading the other users startup folder; added user name postfix in log.  - O25 - Fixed rare error in connecting to WMI.  - Updated and improved LoLBin list check mechanism.  - Improved registry functions.  - Corrected free buffer (ILFree => CoTaskMemFree).  - Explicitly allowed expired Microsoft certificates (without timestamp).  - Updated whitelists of O2, O22, O23.  - Added switch /skipErrors	- do not show error messages and prevent warnings and errors from writing to log file at all.  - Added switch /sha256 - calculate SHA256 hash of files.  - Fixed hash progressbar. Tools: > Digital Signature Checker:  - Improved speed, sorting, added columns - Certificate "Valid From", "Valid Until"; exchanged columns "File Name" / "File Path". > Uninstall Manager:  - Fixed "Uninstall application" button is not always worked. > Process Manager:  - "Save" button auto-refreshes process list. Special thanks to Sandor and regist for samples, testing and suggestions.

+++ src/frmEULA.frm
@@ -134,7 +134,7 @@ Private Sub Form_Initialize()
     Dim hModShell   As Long
     Dim pos         As Long
     Dim sTime       As String
-    Dim sCmdLine    As String
+    Dim sCMDLine    As String
     Dim sPath       As String
     Dim sFile       As String
     Dim ExeName     As String
@@ -245,24 +245,24 @@ Private Sub Form_Initialize()
         gNoGUI = True
     End If
     
-    sCmdLine = Replace$(g_sCommandLine, ":", "+")
+    sCMDLine = Replace$(g_sCommandLine, ":", "+")
     
     '/Tool:xxx
-    If Len(sCmdLine) <> 0 Then
-        If InStr(1, sCmdLine, "tool+StartupList", 1) <> 0 Then bRunToolStartupList = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+UninstMan", 1) <> 0 Then bRunToolUninstMan = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+DigiSign", 1) <> 0 Then bRunToolEDS = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+RegUnlocker", 1) <> 0 Then bRunToolRegUnlocker = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+ADSSpy", 1) <> 0 Then bRunToolADSSpy = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+Hosts", 1) <> 0 Then bRunToolHosts = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+ProcMan", 1) <> 0 Then bRunToolProcMan = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+CheckLNK", 1) <> 0 Then bRunToolCBL = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+ClearLNK", 1) <> 0 Then bRunToolClearLNK = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+Autoruns", 1) <> 0 Then bRunToolAutoruns = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+Executed", 1) <> 0 Then bRunToolExecuted = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+LastActivity", 1) <> 0 Then bRunToolLastActivity = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+ServiWin", 1) <> 0 Then bRunToolServiWin = True: gNoGUI = True
-        If InStr(1, sCmdLine, "tool+TaskScheduler", 1) <> 0 Then bRunToolTaskScheduler = True: gNoGUI = True
+    If Len(sCMDLine) <> 0 Then
+        If InStr(1, sCMDLine, "tool+StartupList", 1) <> 0 Then bRunToolStartupList = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+UninstMan", 1) <> 0 Then bRunToolUninstMan = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+DigiSign", 1) <> 0 Then bRunToolEDS = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+RegUnlocker", 1) <> 0 Then bRunToolRegUnlocker = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+ADSSpy", 1) <> 0 Then bRunToolADSSpy = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+Hosts", 1) <> 0 Then bRunToolHosts = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+ProcMan", 1) <> 0 Then bRunToolProcMan = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+CheckLNK", 1) <> 0 Then bRunToolCBL = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+ClearLNK", 1) <> 0 Then bRunToolClearLNK = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+Autoruns", 1) <> 0 Then bRunToolAutoruns = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+Executed", 1) <> 0 Then bRunToolExecuted = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+LastActivity", 1) <> 0 Then bRunToolLastActivity = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+ServiWin", 1) <> 0 Then bRunToolServiWin = True: gNoGUI = True
+        If InStr(1, sCMDLine, "tool+TaskScheduler", 1) <> 0 Then bRunToolTaskScheduler = True: gNoGUI = True
     End If
     
     ExeName = GetFileName(AppPath(True))
@@ -443,7 +443,10 @@ Sub WatchForProcess()   'waiting for process completion to release unpacked reso
             CloseHandle hProc
         End If
         
-        lret = DeleteFileW(StrPtr(sPathComCtl1))
+        If Not FileExists(BuildPath(AppPath(), "_HijackThis.vbp")) Then
+            lret = DeleteFileW(StrPtr(sPathComCtl1))
+        End If
+        
         lret = DeleteFileW(StrPtr(sPathComCtl2))
     End If
